+++ server/app/controllers/admin/AdminProgramBlockQuestionsController.java
@@ -6,9 +6,9 @@ import static views.ViewUtils.ProgramDisplayType.DRAFT;
 import auth.Authorizers.Labels;
 import com.google.common.collect.ImmutableList;
 import controllers.FlashKey;
-import forms.EnumeratorQuestionForm;
 import forms.ProgramQuestionDefinitionOptionalityForm;
-import forms.QuestionFormBuilder;
+import forms.questions.EnumeratorQuestionForm;
+import forms.questions.QuestionFormBuilder;
 import java.util.Map.Entry;
 import java.util.Optional;
 import javax.inject.Inject;
