+++ server/app/controllers/admin/AdminQuestionController.java
@@ -8,12 +8,12 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import controllers.CiviFormController;
 import controllers.FlashKey;
-import forms.EnumeratorQuestionForm;
-import forms.MapFilterForm;
-import forms.MapQuestionForm;
-import forms.MultiOptionQuestionForm;
-import forms.QuestionForm;
-import forms.QuestionFormBuilder;
+import forms.questions.EnumeratorQuestionForm;
+import forms.questions.MapFilterForm;
+import forms.questions.MapQuestionForm;
+import forms.questions.MultiOptionQuestionForm;
+import forms.questions.QuestionForm;
+import forms.questions.QuestionFormBuilder;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Locale;
