+++ server/app/forms/questions/AddressQuestionForm.java
@@ -1,4 +1,4 @@
-package forms;
+package forms.questions;
 
 import services.question.types.AddressQuestionDefinition;
 import services.question.types.QuestionDefinitionBuilder;
