+++ server/app/forms/questions/NumberQuestionForm.java
@@ -1,4 +1,4 @@
-package forms;
+package forms.questions;
 
 import java.util.OptionalLong;
 import services.question.types.NumberQuestionDefinition;
