+++ server/app/forms/questions/PhoneQuestionForm.java
@@ -1,4 +1,4 @@
-package forms;
+package forms.questions;
 
 import services.question.types.PhoneQuestionDefinition;
 import services.question.types.QuestionType;
