+++ server/app/views/admin/programs/ProgramBlocksView.java
@@ -27,8 +27,8 @@ import com.google.inject.Inject;
 import com.google.inject.assistedinject.Assisted;
 import controllers.admin.routes;
 import forms.BlockForm;
-import forms.EnumeratorQuestionForm;
-import forms.QuestionForm;
+import forms.questions.EnumeratorQuestionForm;
+import forms.questions.QuestionForm;
 import j2html.TagCreator;
 import j2html.tags.DomContent;
 import j2html.tags.specialized.ATag;
