+++ server/app/views/admin/questions/MapQuestionSettingsPartialViewModel.java
@@ -2,7 +2,7 @@ package views.admin.questions;
 
 import com.google.common.collect.ImmutableList;
 import controllers.admin.routes;
-import forms.MapQuestionForm;
+import forms.questions.MapQuestionForm;
 import java.util.List;
 import java.util.OptionalInt;
 import lombok.Builder;
