+++ server/app/views/admin/questions/QuestionConfig.java
@@ -14,15 +14,15 @@ import static j2html.TagCreator.text;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
-import forms.AddressQuestionForm;
-import forms.DateQuestionForm;
-import forms.EnumeratorQuestionForm;
-import forms.FileUploadQuestionForm;
-import forms.IdQuestionForm;
-import forms.MultiOptionQuestionForm;
-import forms.NumberQuestionForm;
-import forms.QuestionForm;
-import forms.TextQuestionForm;
+import forms.questions.AddressQuestionForm;
+import forms.questions.DateQuestionForm;
+import forms.questions.EnumeratorQuestionForm;
+import forms.questions.FileUploadQuestionForm;
+import forms.questions.IdQuestionForm;
+import forms.questions.MultiOptionQuestionForm;
+import forms.questions.NumberQuestionForm;
+import forms.questions.QuestionForm;
+import forms.questions.TextQuestionForm;
 import j2html.tags.specialized.ButtonTag;
 import j2html.tags.specialized.DivTag;
 import j2html.tags.specialized.FieldsetTag;
