+++ server/test/forms/DateQuestionFormTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static services.question.types.DateQuestionDefinition.DateValidationOption.DateType.ANY;
 import static services.question.types.DateQuestionDefinition.DateValidationOption.DateType.APPLICATION_DATE;
 
+import forms.questions.DateQuestionForm;
 import java.util.Locale;
 import java.util.Optional;
 import java.util.UUID;
