+++ server/test/forms/NumberQuestionFormTest.java
@@ -2,6 +2,7 @@ package forms;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import forms.questions.NumberQuestionForm;
 import java.util.Locale;
 import java.util.UUID;
 import org.junit.Test;
